// assets/ts/Interfaces/iSetMethods.ts
export default interface ISetMethods {
    id: (value:any) => void;
    image:(value:any) => void;
    name: (value:any) => void;
    desc: (value:any) => void;
    keys: (value:any) => void;
    stars: (value:any) => void;
    size: (value:any) => void;
    type: (value:any) => void;
    height: (value:any) => void;
    width: (value:any) => void;
}